PLATUI-1451: Updated accessibilit linter and restructured test directories
