+++ project/plugins.sbt
@@ -10,4 +10,4 @@ addSbtPlugin("uk.gov.hmrc"       % "sbt-distributables"       % "2.1.0")
 addSbtPlugin("com.typesafe.play" % "sbt-plugin"               % "2.8.7")
 addSbtPlugin("com.typesafe.sbt"  % "sbt-gzip"                 % "1.0.2")
 addSbtPlugin("org.scalameta"     % "sbt-scalafmt"             % "2.4.0")
-addSbtPlugin("uk.gov.hmrc"       % "sbt-accessibility-linter" % "0.12.0")
+addSbtPlugin("uk.gov.hmrc"       % "sbt-accessibility-linter" % "0.17.0")
